Software and hardware debugged
Software and hardware side of the block analysis fully function, they may still need some tuning but they are mostly correct. Generated a .etl file to find the correct processing time for read and writes on my SSD for this experiment to ensure that time spent in the block layer is estimated to the best of my ability.