Modular Hardware and p95/p99
Made it so the hardware pulls from a file instead of having the commands hardcoded. It has a limit of ~84 commands, but for competency I think running 70 is enough. Also added p95/p99 statistics to the hardware and software, and had both programs output into a file instead of a console. Next step is to generate a python script to graph the differences, and then all that needs to be done is the report and some literature review.