Python Graphing, README, report, and file organization
Added a analysis.py file to generate graphs based on the data generated in software_outputs.txt and hardware_outputs.txt. Added a README to explain where files are stored and how to run the files. Added a report to summarize the findings of the software hardware comparison. Finally, reorganized the file structure to have less files all in the same directory.